+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.0.3
+
+- Jester fix (let me know if that works)
+- Config description
+
 ## v1.0.2
 
 - Flashlight patch